Avoid printing rpmbuild error summary
The expectation is that this change will improve reliability of
autospec's parsing of mock's build.log in the event rpmbuild prints a
large stream of error log messages (e.g. a list of hundreds to thousands
of unpackaged files), since sometimes the error summary becomes
interleaved with the unpackaged file error log messages.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>